Update file structure to match next
