+++ .stylelintrc
@@ -6,9 +6,10 @@
   ],
   "rules": {
     "declaration-empty-line-before": null,
+    "declaration-colon-newline-after": null,
     "keyframes-name-pattern": null,
     "value-keyword-case": ["lower", { ignoreKeywords: ['dummyValue'] }],
     "block-no-empty": null,
-    "number-leading-zero": null
+    "number-leading-zero": null,
   }
 }
