+++ packages/datepickers/src/elements/Datepicker/components/MonthSelector.tsx
@@ -6,7 +6,7 @@
  */
 
 import React, { useCallback } from 'react';
-import { StyledHeader, StyledHeaderPaddle, StyledHeaderLabel } from '../../styled';
+import { StyledHeader, StyledHeaderPaddle, StyledHeaderLabel } from '../../../styled';
 import useDatepickerContext from '../utils/useDatepickerContext';
 
 import ChevronLeftStrokeIcon from '@zendeskgarden/svg-icons/src/16/chevron-left-stroke.svg';
