+++ packages/datepickers/src/elements/DatepickerRange/utils/datepicker-range-reducer.ts
