+++ packages/datepickers/src/styled/StyledMenu.spec.tsx
@@ -7,8 +7,8 @@
 
 import React from 'react';
 import { render, renderRtl } from 'garden-test-utils';
-import { POPPER_PLACEMENT } from '../Datepicker/utils/garden-placements';
-import { StyledMenu } from './styled-menu';
+import { POPPER_PLACEMENT } from '../elements/Datepicker/utils/garden-placements';
+import StyledMenu from './StyledMenu';
 
 describe('StyledMenuView', () => {
   it('renders default styling correctly', () => {
